Add rotating triangle, run everything in main()s
Stupidly I failed to realize that anything in global scope in a module
would run when the module was imported, so each .ts file was doing stuff
to the single svg element, leading to very weird effects.  Nothing
happens till module.main is called now.

Should be using classes for this, probably.